login functionality done, users done

+++ src/views/Chat.vue
@@ -44,7 +44,7 @@ export default {
       theme: 'light',
       showChat: true,
       users: [],
-      currentUserId: '6R0MijpK6M4AIrwaaCY2',
+      currentUserId: '', // sets userid on load
       updatingData: false
     };
   },
@@ -73,20 +73,39 @@ export default {
 
       usersRef.get().then(val => {
         val.forEach(val => {
-          usersRef.doc(val.id).delete();
+          if (val.id !== 'Admin') usersRef.doc(val.id).delete();
+          console.log(val.id);
         });
       });
     },
     async addData() {
+      const userHelper = [
+        {
+          _id: '6R0MijpK6M4AIrwaaCY2',
+          username: 'Luke',
+          avatar: 'https://66.media.tumblr.com/avatar_c6a8eae4303e_512.pnj'
+        },
+        {
+          _id: 'SGmFnBZB4xxMv9V4CVlW',
+          username: 'Leia',
+          avatar: 'https://avatarfiles.alphacoders.com/184/thumb-184913.jpg'
+        },
+        {
+          _id: '6jMsIXUrBHBj7o2cRlau',
+          username: 'Yoda',
+          avatar:
+            'https://vignette.wikia.nocookie.net/teamavatarone/images/4/45/Yoda.jpg/revision/latest?cb=20130224160049'
+        }
+      ];
       this.updatingData = true;
 
-      const user1 = this.users[0];
+      const user1 = userHelper[0];
       await usersRef.doc(user1._id).set(user1);
 
-      const user2 = this.users[1];
+      const user2 = userHelper[1];
       await usersRef.doc(user2._id).set(user2);
 
-      const user3 = this.users[2];
+      const user3 = userHelper[2];
       await usersRef.doc(user3._id).set(user3);
 
       await roomsRef.add({ users: [user1._id, user2._id] });
